+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [1.3.2] - 2025-05-08
+### Added
+- Add customFields option to the plugin to add extra fields globally
+- Added options: {disableText: true } to individual schema to optionally remove text input
 
 ## [1.3.1] - 2025-01-28
 ### Added